tests/inputdevice-script.html

<!DOCTYPE html>
<title>InputDevice test for script-generated objects</title>
<meta name="viewport" content="initial-scale=1">
<meta charset="UTF-8">
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="inputdevice-tests.js"></script>

<h1>InputDevice - test for script-generated objects</h1>
<div id="target"></div>
<div id="log"></div>

<script>
'use strict';
function runTests() {
  test(function() {
    var i = new InputDevice();
    assert_equals(i.firesTouchEvents, false, "firesTouchEvents exists and is false by default.");
    assert_readonly(i, "firesTouchEvents", "firesTouchEvents should be readonly");
  
    i = new InputDevice({});
    assert_equals(i.firesTouchEvents, false, "firesTouchEvents still false with empty init dictionary.");
  
    i = new InputDevice({firesTouchEvents:true});
    assert_equals(i.firesTouchEvents, true, "firesTouchEvents can be initialized to true.");
  
  }, "InputDevice objects can be created correctly from script");
  
  function generateConstructorTest(ctorName) {
    test(function() {
      var ctor = window[ctorName];
      var evt = new ctor("testEvent");
      assert_equals(evt.sourceDevice, null, "sourceDevice exists and is null by default");
      assert_readonly(evt, "sourceDevice", "sourceDevice should be readonly");
      
      var device = new InputDevice({firesTouchEvents:true});
      evt = new ctor("mytest", {sourceDevice: device});
      assert_equals(evt.sourceDevice, device, "sourceDevice can be initialized to a given InputDevice");
      
      var receivedEvent;
      var target = document.getElementById("target");
      target.addEventListener("mytest", function(e) {
        receivedEvent = e;
      });
      target.dispatchEvent(evt);
      assert_equals(receivedEvent.sourceDevice, device, "sourceDevice is preserved across JavaScript event dispatch");
      
    }, "new " + ctorName);
  }

  function generateCreateEventTest(eventType) {
    test(function() {
      var evt = document.createEvent(eventType);
      assert_equals(evt.sourceDevice, null, "sourceDevice exists and is null");
    }, "document.createEvent(\"" + eventType + "\")");
  }


  var eventTypes = ['UIEvent', 'MouseEvent', 'FocusEvent', 'KeyboardEvent', 'WheelEvent', 'CompositionEvent'];
  var optionalEventTypes = ['TouchEvent', 'InputEvent', 'PointerEvent', 'SVGZoomEvent'];
  for( var i = 0; i < optionalEventTypes.length; i++) {
    if (optionalEventTypes[i] in window)
      eventTypes.push(optionalEventTypes[i]);
  }

  for (var i = 0; i < eventTypes.length; i++) {
    var type = eventTypes[i];
    var isFf = navigator.userAgent.indexOf("Firefox") >= 0;

    // TouchEvent is special - support may be disabled even when the TouchEvent exists
    // http://crbug.com/392584
    if (type == "TouchEvent" && !('ontouchstart' in window))
      continue;

    // Firefox doesn't support legacy createEvent for new eventTypes.
    if (!(isFf && ['FocusEvent', 'WheelEvent', 'InputEvent'].indexOf(type) >= 0)) {
      var createEventName = type;
      // createEvent sometimes requires the plural name - highly browser specific
      if (type == "SVGZoomEvent")
        createEventName += "s";
      generateCreateEventTest(createEventName);
    }

    // Firefox doesn't properly support CompositionEvent constructor and there doesn't appear
    // to be any way to feature detect for that. https://bugzilla.mozilla.org/show_bug.cgi?id=1002256.
    if (type == 'CompositionEvent' && isFf)
      continue;

    // Some browsers don't support constructors for all event types.
    if (window[type].length > 0)
      generateConstructorTest(type);
  }
}

</script>
